git-svn-id: svn://123.57.173.36/meijia-parent@153 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ simi-dao/src/main/java/com/simi/po/model/order/Orders.java
@@ -15,7 +15,7 @@ public class Orders {
     
     private Short orderPayType;
 
-    private Short serviceType;
+    private Long serviceType;
     
     private String serviceContent;
 
@@ -81,11 +81,11 @@ public class Orders {
         this.orderNo = orderNo == null ? null : orderNo.trim();
     }
 
-    public Short getServiceType() {
+    public Long getServiceType() {
         return serviceType;
     }
 
-    public void setServiceType(Short serviceType) {
+    public void setServiceType(Long serviceType) {
         this.serviceType = serviceType;
     }
